requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
abseil-cpp=20211102.0=hc377ac9_0
absl-py=1.3.0=py38hca03da5_0
aiohttp=3.8.3=py38h80987f9_0
aiosignal=1.2.0=pyhd3eb1b0_0
appnope=0.1.3=pyhd8ed1ab_0
arrow-cpp=10.0.1=py38h0fd5f32_0
asttokens=2.2.1=pyhd8ed1ab_0
async-timeout=4.0.2=py38hca03da5_0
attrs=22.1.0=py38hca03da5_0
aws-c-common=0.6.8=h1a28f6b_0
aws-c-event-stream=0.1.6=hc377ac9_5
aws-checksums=0.1.11=h1a28f6b_1
aws-sdk-cpp=1.8.185=h4a942e0_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=openblas
blinker=1.4=py38hca03da5_0
boost-cpp=1.73.0=h1a28f6b_12
bottleneck=1.3.5=py38heec5a64_0
brotli=1.0.9=h1a28f6b_7
brotli-bin=1.0.9=h1a28f6b_7
brotlipy=0.7.0=py38h1a28f6b_1002
bzip2=1.0.8=h620ffc9_4
c-ares=1.19.0=h80987f9_0
ca-certificates=2023.01.10=hca03da5_0
cachetools=4.2.2=pyhd3eb1b0_0
catalogue=2.0.7=py38hca03da5_0
certifi=2022.12.7=py38hca03da5_0
cffi=1.15.1=py38h80987f9_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38hca03da5_0
colorama=0.4.6=py38hca03da5_0
cryptography=39.0.1=py38h834c97f_0
cymem=2.0.6=py38hc377ac9_0
cython-blis=0.7.7=py38heec5a64_0
datasets=2.10.1=py38hca03da5_0
debugpy=1.5.1=py38hc377ac9_0
decorator=5.1.1=pyhd8ed1ab_0
dill=0.3.6=py38hca03da5_0
en-core-web-sm=3.3.0=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
ffmpeg=4.2.2=h04105a8_0
filelock=3.9.0=py38hca03da5_0
freetype=2.12.1=h1192e45_0
frozenlist=1.3.3=py38h80987f9_0
fsspec=2023.3.0=py38hca03da5_0
gettext=0.21.0=h13f89a0_1
gflags=2.2.2=hc377ac9_0
giflib=5.2.1=h80987f9_3
glog=0.5.0=hc377ac9_0
gmp=6.2.1=hc377ac9_3
gmpy2=2.1.2=py38h8c48613_0
gnutls=3.6.15=h887c41c_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
grpc-cpp=1.46.1=h8e4afa7_1
grpcio=1.42.0=py38h95c9599_0
huggingface_hub=0.10.1=py38hca03da5_0
icu=68.1=hc377ac9_0
idna=3.4=py38hca03da5_0
importlib-metadata=6.0.0=py38hca03da5_0
ipykernel=6.15.0=pyh736e0ef_0
ipython=8.12.0=pyhd1c38e8_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=py38hca03da5_0
joblib=1.2.0=pypi_0
jpeg=9e=h80987f9_1
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.3.0=py38hca03da5_0
krb5=1.19.4=h8380606_0
lame=3.100=h1a28f6b_0
langcodes=3.3.0=pyhd3eb1b0_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libboost=1.73.0=h49e8a49_12
libbrotlicommon=1.0.9=h1a28f6b_7
libbrotlidec=1.0.9=h1a28f6b_7
libbrotlienc=1.0.9=h1a28f6b_7
libcurl=7.88.1=h0f1d93c_0
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_0
libedit=3.1.20221030=h80987f9_0
libev=4.33=h1a28f6b_1
libevent=2.1.12=hf27765b_0
libffi=3.4.2=hca03da5_6
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h1a28f6b_2
libidn2=2.3.1=h1a28f6b_0
libnghttp2=1.46.0=h95c9599_0
libopenblas=0.3.21=h269037a_0
libopus=1.3=h1a28f6b_1
libpng=1.6.39=h80987f9_0
libprotobuf=3.20.3=h514c7bf_0
libsodium=1.0.18=h27ca646_1
libssh2=1.10.0=hf27765b_0
libtasn1=4.16.0=h1a28f6b_0
libthrift=0.13.0=hd358383_6
libtiff=4.5.0=h313beb8_2
libunistring=0.9.10=h1a28f6b_0
libvpx=1.10.0=hc377ac9_0
libwebp=1.2.4=ha3663a8_1
libwebp-base=1.2.4=h80987f9_1
libxml2=2.10.3=h372ba2a_0
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_0
markdown=3.4.1=py38hca03da5_0
markupsafe=2.1.1=py38h1a28f6b_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mpc=1.1.0=h8c48613_1
mpfr=4.0.2=h695f6f0_1
mpmath=1.2.1=py38hca03da5_0
multidict=6.0.2=py38h1a28f6b_0
multiprocess=0.70.14=py38hca03da5_0
murmurhash=1.0.7=py38hc377ac9_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.7.3=h84b5d62_1
networkx=2.8.4=py38hca03da5_1
nlp=0.4.0=pypi_0
nltk=3.8.1=pypi_0
numexpr=2.8.4=py38h79ee842_0
numpy=1.23.5=py38h1398885_0
numpy-base=1.23.5=py38h90707a3_0
oauthlib=3.2.2=py38hca03da5_0
openh264=1.8.0=h98b2900_0
openssl=1.1.1t=h1a28f6b_0
orc=1.7.4=hdca1487_1
packaging=23.0=pyhd8ed1ab_0
pandas=1.5.3=py38h78102c4_0
parso=0.8.3=pyhd8ed1ab_0
pathy=0.10.1=py38hca03da5_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py38h313beb8_0
pip=23.0.1=py38hca03da5_0
platformdirs=3.2.0=pyhd8ed1ab_0
preshed=3.0.6=py38hc377ac9_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
protobuf=3.20.3=py38h313beb8_0
psutil=5.9.0=py38h1a28f6b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=10.0.1=py38h23c13bc_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.8.2=pypi_0
pygments=2.14.0=pyhd8ed1ab_0
pyjwt=2.4.0=py38hca03da5_0
pyopenssl=23.0.0=py38hca03da5_0
pysocks=1.7.1=py38hca03da5_0
python=3.8.15=hc0d8a6c_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python-xxhash=2.0.2=py38h1a28f6b_1
python_abi=3.8=2_cp38
pytorch=2.0.0=py3.8_0
pytz=2022.7=py38hca03da5_0
pyyaml=6.0=py38h80987f9_1
pyzmq=23.2.0=py38hc377ac9_0
re2=2022.04.01=hc377ac9_0
readline=8.2=h1a28f6b_0
regex=2023.3.23=pypi_0
requests=2.28.1=py38hca03da5_1
requests-oauthlib=1.3.0=py_0
responses=0.13.3=pyhd3eb1b0_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
sentencepiece=0.1.98=pypi_0
setuptools=65.6.3=py38hca03da5_0
shellingham=1.5.0=py38hca03da5_0
six=1.16.0=pyh6c4a22f_0
smart_open=5.2.1=py38hca03da5_0
snappy=1.1.9=hc377ac9_0
spacy=3.3.1=py38h5c6307a_0
spacy-legacy=3.0.9=py38hca03da5_0
spacy-loggers=1.0.1=pyhd3eb1b0_0
sqlite=3.41.1=h80987f9_0
srsly=2.4.6=py38h313beb8_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.11.1=py38hca03da5_0
tensorboard=2.11.0=py38hca03da5_0
tensorboard-data-server=0.6.1=py38ha6e5c4f_0
tensorboard-plugin-wit=1.6.0=py_0
thinc=8.0.15=py38h5c6307a_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=hb8d0fd4_0
torch-tb-profiler=0.4.1=pypi_0
torchaudio=2.0.0=py38_cpu
torchdata=0.6.0=pypi_0
torchtext=0.6.0=pypi_0
torchvision=0.15.0=py38_cpu
tornado=6.1=py38hea4295b_1
tqdm=4.65.0=py38h86d0a89_0
traitlets=5.9.0=pyhd8ed1ab_0
typer=0.4.1=py38hca03da5_0
typing-extensions=4.5.0=hd8ed1ab_0
typing_extensions=4.5.0=pyha770c72_0
urllib3=1.26.15=py38hca03da5_0
utf8proc=2.6.1=h1a28f6b_0
wasabi=0.9.1=py38hca03da5_0
wcwidth=0.2.6=pyhd8ed1ab_0
werkzeug=2.2.3=py38hca03da5_0
wheel=0.38.4=py38hca03da5_0
x264=1!152.20180806=h1a28f6b_0
xxhash=0.8.0=h1a28f6b_3
xz=5.2.10=h80987f9_1
yaml=0.2.5=h1a28f6b_0
yarl=1.8.1=py38h1a28f6b_0
zeromq=4.3.4=hbdafb3b_1
zipp=3.11.0=py38hca03da5_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.4=hd90d995_0
